Add program flow + Python API file
I think we will need to create a python file to write the API.
We will be able to select a some points and receive the organized array.
The new configuration of thw weather data database does not allow us to
receive the data as we used to do.
The data.py file is not working yet, but I will try to talk to Apolo to
see if he could help us.